+++ langworld_db_data/main.py
@@ -1,7 +1,7 @@
 from pycldf import Dataset
 
-from langworld_db_data.cldfwriters.cldf_dataset_writer import CLDFDatasetWriter
 from langworld_db_data.constants.paths import DATA_DIR, FILE_WITH_CLDF_DATASET_METADATA
+from langworld_db_data.export.cldf_dataset_writer import CLDFDatasetWriter
 from langworld_db_data.mdlisters.custom_value_lister import CustomValueLister
 from langworld_db_data.mdlisters.listed_value_lister import ListedValueLister
 from langworld_db_data.tools.common.files.csv_xls import check_csv_for_malformed_rows
